requirements.txt
requests
colorama
selenium
keyboard
rapidfuzz
pandas
python-dotenv